UPSTREAM: tss: implement tlcl_save_state
When an untrusted OS is running, we would like to use the Cr50
vendor-specific VENDOR_CC_TPM_MODE command to disable TPM.
Before doing this, we should save TPM state.  Implement
tlcl_save_state for this purpose.

This needs to live in coreboot codebase since on S3 resume path,
depthcharge is not reached.

Implement the function in both tcg-1.2 and tcg-2.0 for
completeness.

BUG=b:70681930,b:118202153
TEST=hack a call to tlcl_save_state into coreboot on S3 resume
     verify in AP console that it is called

Change-Id: Ide33d036354072fd267a82b10ef8fa3c4c451df4
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2e690eeaf2c59070f74b01afb238c8a5208228f0
Original-Signed-off-by: Joel Kitching <kitching@google.com>
Original-Change-Id: I8b51ca68456fc9b655e4dc2d0958b7c040d50510
Original-Reviewed-on: https://review.coreboot.org/c/29646
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Philipp Deppenwiese <zaolin.daisuki@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/1355083
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>