UPSTREAM: mb/google/sarien: Set SMBIOS mainboard SKU
Setting sku_id() is not enough to get a value to show up in the SMBIOS
tables, it also needs to be returned as a string for the table creation
to consume.  This change defines the smbios_mainboard_sku() function
and returns a string constant of "sku#" as expected.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ifcda305ea27dd7ab9efc3302f7e5898c4795ce7f
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: bf2710e8496327a9e9a1f5b4517ae2676ba78612
Original-Change-Id: I03013bab89d53d1eba969c6ffb7e95fcbb315a81
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/29649
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Trent Begin <tbegin@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1341402
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>