UPSTREAM: ec/google/wilco/acpi: Add DPTF support
Add the support needed for DPTF.  This includes the methods to
write trip point values, read temperatures, and handle events.

This was tested on a sarien board by inspecting AML debug output
with the kernel while monitoring temperatures and trip points in
sysfs and controlling temperatures with a fan to ensure that when
a trip point is crossed an SCI is generated and the event is
handled properly.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic13053f4aaa5abcc77ce2fa3236b393abda1c935
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 98d7de7ea93379957fc3f48bef6912e9947e1099
Original-Change-Id: I8d8570d176c0896fa709a6c782b319f58d3c1e52
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/29761
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1363372
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>