UPSTREAM: intelblocks/cpu: Add function to set CPU clock to minimum value
Provide a library function to set the CPU frequency to minimum
value. This will result in the lowest possible CPU clock with
the lowest possible power consumption. This can be useful in mobile
devices where the power dissipation is limited.

This setting can be overruled by the OS if it has an p-state driver
which can adjust the clock to it's need.

BUG=none
BRANCH=none
TEST=none

Change-Id: I5fdb18f7de214258cb48ec849aadfe6879841b61
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 52c58929fd8d9011e7990c042418624b567ffa1b
Original-Change-Id: I817095b13ab8cbaab82f25c72947b00ee854d549
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/29771
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Reviewed-on: https://chromium-review.googlesource.com/1350144
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>