UPSTREAM: soc/intel/apollolake: Add Kconfig switch to enable minimum clock ratio
Add a Kconfig switch to be able to set the CPU clock to the lowest
possible ratio. If enabled the CPU will consume as little power as
possible while providing the lowest performance.

This setting can be overruled by the OS if it has an p-state driver
which can adjust the clock to its need.

BUG=none
BRANCH=none
TEST=none

Change-Id: I9cb361badcbbcd3bbf3b75c1448010f91fa3025d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 26361862bdff9bc12c31e4f153e83422e21179b6
Original-Change-Id: I4a59586da72d1915749110a36f565fe2aa69e073
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/29772
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/1350145
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>