UPSTREAM: src/mb/pcengines/apu2/mainboard.c: Fix retrieving serial number
Handle situation when first NIC is not BDF 1:0.0. The PCI enumeration
is different when a external PCIe device is connected to mPCIe2 slot
which is routed to first PCIe bridge. The first NIC is then assigned
BDF 2:0.0, because it is connected to the second PCIe bridge.

Read the secondary bus number from the NIC PCIe bridge before attempting
to read MAC adress and calculating serial number.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic1d8a97db54cc329b032ea688356810aece3fd37
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 0486458c730a2424ce2decc43dbea7610eb09957
Original-Change-Id: I9f89a6f3cd0c23a2d2924e587338f69c260b12f8
Original-Signed-off-by: Micha ygowski <michal.zygowski@3mdeb.com>
Original-Reviewed-on: https://review.coreboot.org/c/29842
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/1358825
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>