UPSTREAM: src/(device/lib/soc): Remove unused variables
When building grunt with flags set to detect variables that get a value but
then are unused, there are 5 instances that causes error (unused variable).
In most cases it's enough to simply remove the variable. Other instances,
is better to simply use the variables (one instance it's a return value, on
the other instance using the variables makes code more readable).

BUG=b:120260448
TEST=Build and boot grunt.

Change-Id: Id755dae671dc0b09dbf7b996d271cf782c3e4352
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6a9e6cd89e2be6b35675ee81c2272a5ea5f2db75
Original-Change-Id: I0d00fb6a42db20afb34c76b9445a741a57096ead
Original-Signed-off-by: Richard Spiegel <richard.spiegel@silverbackltd.com>
Original-Reviewed-on: https://review.coreboot.org/c/29985
Original-Reviewed-by: HAOUAS Elyes <ehaouas@noos.fr>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1363402
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>