UPSTREAM: acpi_pld: Make it easier to define the ACPI USB device groups
The Linux kernel can use the ACPI _PLD group information to
determine peer ports.  Currently to define the group information
the devicetree must provide a complete _PLD structure.  This
change pulls the group information into a separate structure that
can be defined in devicetree.  This makes it easier to set for
USB devices in devicetree that do not need a full custom PLD.

This was tested on a sarien board with the USB devices defined
by verifying that the USB 2/3 ports are correctly identified
with their peer in sysfs.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic895b8d44bf38d2de7dcd863357b4939a25dfab1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e1eca1d91cf05b8769e830b5ab59060a0ad5ff0c
Original-Change-Id: Ifd4cadf0f6c901eb3832ad4e1395904f99c2f5a0
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/29998
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1363378
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>