UPSTREAM: Documentation/flash_tutorial/index.md: warn about dots painted on ICs
I fried my mainboard because I tried to orient my chip by lining a blue dot on
the corner of my chip with a dot depicted on the chip datasheet.  They
apparently have nothing to do with each other, and this is normal.  Add
warning about this to the docs to hopefully spare others from a similar fate.

BUG=none
BRANCH=none
TEST=none

Change-Id: I985f3514a83af65fe9722b61372b342076fdd120
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8bed5efad7d9444fdf02665eac3ad84efb35d42e
Original-Signed-off-by: Michael Bacarella <michael.bacarella@gmail.com>
Original-Change-Id: Ib634589aaa11f75bde2ef2e13d2cacc4cae19a3f
Original-Signed-off-by: Michael Bacarella <michael.bacarella@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/30028
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://chromium-review.googlesource.com/1363409
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>