UPSTREAM: ec/google/wilco: Guard DTPF with ifdef
There is a dependency issue with the EC DPTF code accessing
methods that are external, but once the mainboard includes the
relevant code they become internal and the current version of
IASL used by jenkins will fail to compile it.

Until the new IASL is deployed everywhere wrap the EC DPTF code
and expect that the mainboard will explicitly enable it.

BUG=none
BRANCH=none
TEST=none

Change-Id: Idacd7a9713aa3727f4a0b4689525d5de2360a383
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: bfb001d1a0b953fd832fce6bb13c02657ab1139f
Original-Change-Id: I612ad8f86d424060ca0303d267d7c2915c760173
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/30036
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1363373
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>