UPSTREAM: mb/google/sarien: Update GPIOs for next build
Update the GPIOs for the next board build.  Mostly minor changes but
the polarity change on GPP_E8/RECOVERY on sarien will result in it
booting to recovery every time unless using new hardware.

For this reason the recovery mode GPIO that is passed to vboot is
commented out for sarien.  It is only used for testing and currently
it is useful to have an image that works on both board versions.

BUG=none
BRANCH=none
TEST=none

Change-Id: I03781dc1ebca904c0a2c54cba2873f085fa90c13
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 025a03f6164b4f24e36aea39688b39abdec28fee
Original-Change-Id: I32d84f3010cb4d3968370a03f7e191b1710a50e8
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/30062
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1370269
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>