UPSTREAM: drivers/i2c/designware: Add soc_clock entry for 216MHz
Add an entry to the soc_clock table for a 216MHz clock so that
the I2C controller clock is calculated correctly when the I2C
bus is used in coreboot.

This was tested by measuring the I2C clock speed on H1 I2C bus
on a sarien board in coreboot and ensuring it is ~400KHz.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ifce2bf096c85c3a866567fb58cf81768a19718c2
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 4afefd648b28d4ebabd6d07790037e3e00c9ff03
Original-Change-Id: I6c3cacdad318a5ce41bc41e3ac81385c2d4f396c
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/30068
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1367630
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>