UPSTREAM: mb/google/octopus/bobba: Add support to handle PEN_EJECT event
Enable gpio_keys driver for bobba and add required configuration in the
device tree to handle the pen eject event.

BRANCH=octopus
BUG=b:117953118
TEST=Ensure that the system boots to ChromeOS. Ensure that the stylus tools
open on pen eject. Ensure that the system wakes on Pen Eject. Ensure that
the system enters S0ix and S3 states after the pen is ejected. Ensure that
the system enters S0ix and S3 states when the pen remains inserted in its
holder. Ensured that the system does not wake when the pen is inserted.
Ensure that the suspend_stress_test runs successfully for 25 iterations
with the pen placed in its holder.

Change-Id: I87986a7cffb6f0aed9deb8de90fc0bd887663299
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2b27e236b52fc3ce022bae7bf0eafb7651a86554
Original-Change-Id: I768b89d2b45f4dcab6d235b11ce00544a827f22d
Original-Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/30108
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1434503
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>