UPSTREAM: Revert "mb/google/poppy/variants/nocturne: Add DMIC properties to ACPI DSD"
This reverts commit 999b916015ea0558e3821bdb51501b43a60b5ed6.

The DMIC doesn't have an ACPI id. The patch which enables ACPI
device with id DMIC may create conflict in the feature. Also the
ACPI id "DMIC"  doesn't comply with ACPI naming conventions. The
issue for which the patch was introduced, is already addressed in
kernel DMIC driver and the patches are upstreamed in to the Linux
kernel.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia3cebba6511de3d3793b86426c8e7247d0309887
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 096833fef4eb5616a95559bb2931dc5e8c0cf594
Original-Change-Id: I42cb076700dcb5906599471bebfcd5b265b17644
Original-Signed-off-by: Jenny TC <jenny.tc@intel.com>
Original-Reviewed-on: https://review.coreboot.org/c/30151
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1389180
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>