UPSTREAM: intelblocks/systemagent: Add ACPI table generation hook
In preparation of generating DMAR tables, provide the hook in SoC scope
for the systemagent to write ACPI tables. The complete functionality is
SoC-specific. Therefore the entry hook is defined as a weak function which
can be overridden by SoC code. If the SoC does not have support for
generating DMAR tables this hook will do no harm.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ifda05f3ee0c8bd118b6186e5c53243835674b68b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d12530cb838c4fa2d503d27c9ea19e8ffb515725
Original-Change-Id: I1333ae2b79f1a855e6f3bb39bf534da170ddc9e1
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/30989
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/1439226
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>