UPSTREAM: nb/intel/haswell: Handle boards that do not support IGD
Processor graphics is disabled on, for example, the C222 and C224
chipsets.

The change to resource assignment in northbridge.c prevents the following
warning that occurs when the IGD is disabled:

> skipping PCI: 00:00.0@3 fixed resource, size=0!

Tested on a Supermicro X10SLM+-F, which has the IGD disabled by the
chipset. The graphics memory is reclaimed and no issues were observed.

Also tested on an ASRock H81M-HDS. This board has an IGD, but no
regressions were observed.

BUG=none
BRANCH=none
TEST=none

Change-Id: I100ddc7f40757be79c6e21f57d49d5a1e7051414
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c5d367b4d3adbe50a91029710bce69c64b5fa3df
Original-Change-Id: I86d4aef50b6588f08b86c9758a4b95ccd65e9a96
Original-Signed-off-by: Tristan Corrick <tristan@corrick.kiwi>
Original-Reviewed-on: https://review.coreboot.org/c/30271
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/1392659
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>